+++ spec/lib/ruby-lokalise-api/rest/projects_spec.rb
@@ -22,6 +22,11 @@ RSpec.describe Lokalise::Client do
       expect(projects.total_pages).to eq(5)
       expect(projects.results_per_page).to eq(1)
       expect(projects.current_page).to eq(2)
+
+      expect(projects.next_page?).to eq(true)
+      expect(projects.last_page?).to eq(false)
+      expect(projects.prev_page?).to eq(true)
+      expect(projects.first_page?).to eq(false)
     end
   end