+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.1.0
+
+* Added methods to work with pagination (`next_page?`, `last_page?`, `prev_page?`, `first_page?`, `next_page`, `prev_page`)
+
+## 1.0.1 (10-Dec-18)
+
+* Fixed incorrect build
+
 ## 1.0.0 (10-Dec-18)
 
 * Initial release